Task #2: Support hours in task duration (fractional days)
- Schema: Changed durationDays from integer to numeric in tasks,
  importedProjectTasks, templateTasks tables (shared/schema.ts)
- Utilities: Added formatDuration, parseDurationInput, hoursToFractionalDays,
  fractionalDaysToHours to workingDays.ts; HOURS_PER_DAY=8 constant
- UI: Updated all duration inputs to text fields accepting "4h", "2d", "1d 4h"
  format in CreateTaskDialog, ProjectTasksTab, Tasks.tsx, ProjectGanttView
- Display: All duration displays use formatDuration() instead of "${x}d"
  across ProjectGanttView, Tasks.tsx, Templates.tsx, ProjectDetails.tsx
- Import parsers: MS Project XML now preserves fractional hours (PT4H→0.5d),
  Excel parser handles "4 hours", "1.5 days" etc., CSV import uses parseFloat
- Dataverse: Fixed duration conversion from minutes to working days (÷480)
- Server: Updated calculateEndDate to handle fractional days via Math.floor
- Gantt inline edit: Added null guard to prevent invalid input coercion to 0

+++ shared/schema.ts
@@ -429,7 +429,7 @@ export const tasks = pgTable("tasks", {
   baselineEndDate: date("baseline_end_date"), // Original planned end
   actualStartDate: date("actual_start_date"), // When work actually started
   actualEndDate: date("actual_end_date"), // When work actually finished
-  durationDays: integer("duration_days"), // Duration in days - auto-calculates endDate if set
+  durationDays: numeric("duration_days"), // Duration in days (supports fractional, e.g. 0.5 = 4h)
   estimatedHours: numeric("estimated_hours"), // Estimated effort in hours
   actualHours: numeric("actual_hours"), // Actual hours worked
   remainingHours: numeric("remaining_hours"), // Remaining effort
@@ -1269,7 +1269,7 @@ export const mppImportTasks = pgTable("mpp_import_tasks", {
   startDate: date("start_date"),
   finishDate: date("finish_date"),
   duration: text("duration"), // Duration as text (e.g., "5 days")
-  durationDays: integer("duration_days"), // Duration in days
+  durationDays: numeric("duration_days"), // Duration in days (supports fractional)
   percentComplete: integer("percent_complete").default(0),
   outlineLevel: integer("outline_level").default(1), // Task hierarchy level
   parentTaskId: integer("parent_task_id"), // Parent task reference
@@ -2759,7 +2759,7 @@ export const projectTemplateItems = pgTable("project_template_items", {
   startDate: date("start_date"),
   endDate: date("end_date"),
   duration: text("duration"),
-  durationDays: integer("duration_days"),
+  durationDays: numeric("duration_days"),
   outlineLevel: integer("outline_level").default(1),
   parentTaskId: integer("parent_task_id"),
   isSummary: boolean("is_summary").default(false),
